// WebApp.Client/src/generated/dtos.ts
export class ProblemDetails implements IProblemDetails {
    type?: string | undefined;
    title?: string | undefined;
    status?: number | undefined;
    detail?: string | undefined;
    instance?: string | undefined;

    [key: string]: any;

    constructor(data?: IProblemDetails) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            for (var property in _data) {
                if (_data.hasOwnProperty(property))
                    this[property] = _data[property];
            }
            this.type = _data["type"];
            this.title = _data["title"];
            this.status = _data["status"];
            this.detail = _data["detail"];
            this.instance = _data["instance"];
        }
    }

    static fromJS(data: any): ProblemDetails {
        data = typeof data === 'object' ? data : {};
        let result = new ProblemDetails();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        for (var property in this) {
            if (this.hasOwnProperty(property))
                data[property] = this[property];
        }
        data["type"] = this.type;
        data["title"] = this.title;
        data["status"] = this.status;
        data["detail"] = this.detail;
        data["instance"] = this.instance;
        return data;
    }
}

export interface IProblemDetails {
    type?: string | undefined;
    title?: string | undefined;
    status?: number | undefined;
    detail?: string | undefined;
    instance?: string | undefined;

    [key: string]: any;
}
// ...
export class CategoryHeaderDto implements ICategoryHeaderDto {
    id?: number;
    name?: string;
    description?: string;
    icon?: string;
    color?: string;
    isDefault?: boolean;

    constructor(data?: ICategoryHeaderDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.description = _data["description"];
            this.icon = _data["icon"];
            this.color = _data["color"];
            this.isDefault = _data["isDefault"];
        }
    }

    static fromJS(data: any): CategoryHeaderDto {
        data = typeof data === 'object' ? data : {};
        let result = new CategoryHeaderDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["description"] = this.description;
        data["icon"] = this.icon;
        data["color"] = this.color;
        data["isDefault"] = this.isDefault;
        return data;
    }
}

export interface ICategoryHeaderDto {
    id?: number;
    name?: string;
    description?: string;
    icon?: string;
    color?: string;
    isDefault?: boolean;
}
// ...
export class TransactionHeaderDto implements ITransactionHeaderDto {
    id!: number;
    name!: string;
    description!: string;
    amount!: number;
    date!: Date;
    createdAt!: Date;
    createdByUser!: UserMinimalDto;
    category!: CategoryHeaderDto | undefined;

    constructor(data?: ITransactionHeaderDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
        if (!data) {
            this.createdByUser = new UserMinimalDto();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.description = _data["description"];
            this.amount = _data["amount"];
            this.date = _data["date"] ? new Date(_data["date"].toString()) : undefined as any;
            this.createdAt = _data["createdAt"] ? new Date(_data["createdAt"].toString()) : undefined as any;
            this.createdByUser = _data["createdByUser"] ? UserMinimalDto.fromJS(_data["createdByUser"]) : new UserMinimalDto();
            this.category = _data["category"] ? CategoryHeaderDto.fromJS(_data["category"]) : undefined as any;
        }
    }

    static fromJS(data: any): TransactionHeaderDto {
        data = typeof data === 'object' ? data : {};
        let result = new TransactionHeaderDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["description"] = this.description;
        data["amount"] = this.amount;
        data["date"] = this.date ? this.date.toISOString() : undefined as any;
        data["createdAt"] = this.createdAt ? this.createdAt.toISOString() : undefined as any;
        data["createdByUser"] = this.createdByUser ? this.createdByUser.toJSON() : undefined as any;
        data["category"] = this.category ? this.category.toJSON() : undefined as any;
        return data;
    }
}

export interface ITransactionHeaderDto {
    id: number;
    name: string;
    description: string;
    amount: number;
    date: Date;
    createdAt: Date;
    createdByUser: UserMinimalDto;
    category: CategoryHeaderDto | undefined;
}

export class TransactionDataDto extends TransactionHeaderDto implements ITransactionDataDto {
    wallet!: WalletHeaderDto;

    constructor(data?: ITransactionDataDto) {
        super(data);
        if (!data) {
            this.wallet = new WalletHeaderDto();
        }
    }

    override init(_data?: any) {
        super.init(_data);
        if (_data) {
            this.wallet = _data["wallet"] ? WalletHeaderDto.fromJS(_data["wallet"]) : new WalletHeaderDto();
        }
    }

    static override fromJS(data: any): TransactionDataDto {
        data = typeof data === 'object' ? data : {};
        let result = new TransactionDataDto();
        result.init(data);
        return result;
    }

    override toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["wallet"] = this.wallet ? this.wallet.toJSON() : undefined as any;
        super.toJSON(data);
        return data;
    }
}

export interface ITransactionDataDto extends ITransactionHeaderDto {
    wallet: WalletHeaderDto;
}

export class WalletHeaderDto implements IWalletHeaderDto {
    id?: number;
    name?: string;
    balance?: number;
    createdAt?: Date;
    updatedAt?: Date;

    constructor(data?: IWalletHeaderDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.balance = _data["balance"];
            this.createdAt = _data["createdAt"] ? new Date(_data["createdAt"].toString()) : undefined as any;
            this.updatedAt = _data["updatedAt"] ? new Date(_data["updatedAt"].toString()) : undefined as any;
        }
    }

    static fromJS(data: any): WalletHeaderDto {
        data = typeof data === 'object' ? data : {};
        let result = new WalletHeaderDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["balance"] = this.balance;
        data["createdAt"] = this.createdAt ? this.createdAt.toISOString() : undefined as any;
        data["updatedAt"] = this.updatedAt ? this.updatedAt.toISOString() : undefined as any;
        return data;
    }
}

export interface IWalletHeaderDto {
    id?: number;
    name?: string;
    balance?: number;
    createdAt?: Date;
    updatedAt?: Date;
}
// ...
export class WalletDataDto extends WalletHeaderDto implements IWalletDataDto {
    users?: WalletUserDto[];
    transactions?: TransactionHeaderDto[];

    constructor(data?: IWalletDataDto) {
        super(data);
    }

    override init(_data?: any) {
        super.init(_data);
        if (_data) {
            if (Array.isArray(_data["users"])) {
                this.users = [] as any;
                for (let item of _data["users"])
                    this.users!.push(WalletUserDto.fromJS(item));
            }
            if (Array.isArray(_data["transactions"])) {
                this.transactions = [] as any;
                for (let item of _data["transactions"])
                    this.transactions!.push(TransactionHeaderDto.fromJS(item));
            }
        }
    }

    static override fromJS(data: any): WalletDataDto {
        data = typeof data === 'object' ? data : {};
        let result = new WalletDataDto();
        result.init(data);
        return result;
    }

    override toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.users)) {
            data["users"] = [];
            for (let item of this.users)
                data["users"].push(item ? item.toJSON() : undefined as any);
        }
        if (Array.isArray(this.transactions)) {
            data["transactions"] = [];
            for (let item of this.transactions)
                data["transactions"].push(item ? item.toJSON() : undefined as any);
        }
        super.toJSON(data);
        return data;
    }
}

export interface IWalletDataDto extends IWalletHeaderDto {
    users?: WalletUserDto[];
    transactions?: TransactionHeaderDto[];
}
// ...
export class CreateTransactionDto implements ICreateTransactionDto {
    walletId?: number | undefined;
    categoryId?: number | undefined;
    name!: string;
    description?: string | undefined;
    amount!: number;
    date!: Date;

    constructor(data?: ICreateTransactionDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.walletId = _data["walletId"];
            this.categoryId = _data["categoryId"];
            this.name = _data["name"];
            this.description = _data["description"];
            this.amount = _data["amount"];
            this.date = _data["date"] ? new Date(_data["date"].toString()) : undefined as any;
        }
    }

    static fromJS(data: any): CreateTransactionDto {
        data = typeof data === 'object' ? data : {};
        let result = new CreateTransactionDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["walletId"] = this.walletId;
        data["categoryId"] = this.categoryId;
        data["name"] = this.name;
        data["description"] = this.description;
        data["amount"] = this.amount;
        data["date"] = this.date ? this.date.toISOString() : undefined as any;
        return data;
    }
}

export interface ICreateTransactionDto {
    walletId?: number | undefined;
    categoryId?: number | undefined;
    name: string;
    description?: string | undefined;
    amount: number;
    date: Date;
}